api/pre_project_spikes/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
elastic-transport==8.17.0
elasticsearch==8.17.1
ibm-cloud-sdk-core==3.22.1
ibm-watson==9.0.0
idna==3.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
websocket-client==1.8.0
bs4==4.13.3